requirements.txt
openai
python-dotenv
python-telegram-bot==13.13
urllib3
